Alternative way of rendering line/token/char spans - unfinished
