+++ src/components/TypingExercise/LineSpan/LineSpan.css
@@ -1,0 +1,3 @@
+.line {
+  display: flex;
+}
