--- src/components/TypingExercise/Statistics/Acuracy/Acuracy.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Acuracy = () => {
-  return (
-    <div>
-      
-    </div>
-  );
-};
-
-export default Acuracy;
